agregar botones de edicion y reubicacion de la fun
