+++ src/components/Feed.js
@@ -1,6 +1,10 @@
 import {
-  getUser, savePost, showPosts, addPostToFeed,
+  doc, onSnapshot,
+} from 'firebase/firestore';
+import {
+  getUser, savePost, showPosts,
 } from '../firebase/firebase.js';
+import { db } from '../firebase/config.js';
 
 export const Feed = () => {
   // ---------------------------HEAD----------------------
@@ -264,6 +268,7 @@ export const Feed = () => {
 
   // Feed container
   const feedContainer = document.createElement('div');
+  feedContainer.className = 'timeline-container';
 
   // --------------------------INICIO---Footer
   const footerFeed = document.createElement('footer');
@@ -334,6 +339,89 @@ export const Feed = () => {
   flexItemRightDiv.appendChild(mainElement);
   centeredMainDiv.appendChild(flexItemLeftDiv);
   centeredMainDiv.appendChild(flexItemRightDiv);
+
+  // --------------------------------------------------------------
+  // INICIO DE FUNCIONES A BOTONES E INTERACCIONES EN FEED
+  // --------------------------------------------------------------
+
+  // Creacion de un post
+  function addPostToFeed(id, content, container, name) {
+    const userPublishedPost = document.createElement('div');
+    userPublishedPost.className = 'user-published-post';
+
+    const userPublishedPostDiv = document.createElement('div');
+    userPublishedPostDiv.className = 'user-published-post-div';
+    userPublishedPost.appendChild(userPublishedPostDiv);
+
+    const userPublishedPostContent = document.createElement('div');
+    userPublishedPostContent.className = 'user-published-post-content';
+    userPublishedPostDiv.appendChild(userPublishedPostContent);
+
+    const conversationImg3 = document.createElement('img');
+    conversationImg3.className = 'conversation-img colorlightblue';
+    conversationImg3.src = 'img/istockphoto-1323400501-612x612.jpg';
+    userPublishedPostContent.appendChild(conversationImg3);
+
+    const userPublishedPostTextContent = document.createElement('div');
+    userPublishedPostTextContent.className = 'user-published-post-text-content';
+    userPublishedPostContent.appendChild(userPublishedPostTextContent);
+
+    const textContentUpperRow = document.createElement('div');
+    textContentUpperRow.className = 'text-content-upper-row';
+    userPublishedPostTextContent.appendChild(textContentUpperRow);
+
+    const userPublishedPostTitle = document.createElement('p');
+    userPublishedPostTitle.className = 'user-published-post-title';
+    userPublishedPostTitle.textContent = name;
+    textContentUpperRow.appendChild(userPublishedPostTitle);
+
+    const userPublishedPostEdit = document.createElement('p');
+    userPublishedPostEdit.className = 'user-published-post-edit';
+    userPublishedPostEdit.textContent = '...';
+    textContentUpperRow.appendChild(userPublishedPostEdit);
+    //
+    const PostEditButtons = document.createElement('ul');
+    PostEditButtons.className = 'edit-post-option-container colorwhite';
+    userPublishedPostEdit.appendChild(PostEditButtons);
+
+    const editButton = document.createElement('li');
+    editButton.className = 'edit-post-option';
+    editButton.textContent = 'editar';
+    PostEditButtons.appendChild(editButton);
+
+    const eraseButton = document.createElement('li');
+    eraseButton.className = 'edit-post-option two';
+    eraseButton.textContent = 'eliminar';
+    PostEditButtons.appendChild(eraseButton);
+    //
+    const userPublishedPostText = document.createElement('p');
+    userPublishedPostText.className = 'user-published-post-text';
+    userPublishedPostText.textContent = content;
+    userPublishedPostTextContent.appendChild(userPublishedPostText);
+
+    const userPublishedPostActions = document.createElement('div');
+    userPublishedPostActions.className = 'user-published-post-actions';
+    userPublishedPostDiv.appendChild(userPublishedPostActions);
+
+    const messagesIcon3 = document.createElement('img');
+    messagesIcon3.className = 'messages-icon';
+    messagesIcon3.src = 'img/chat-svgrepo-com.svg';
+    messagesIcon3.alt = 'message-icon-for-comment';
+    userPublishedPostActions.appendChild(messagesIcon3);
+
+    const heartIcon3 = document.createElement('img');
+    heartIcon3.className = 'heart-icon';
+    heartIcon3.src = 'img/heart-svgrepo-com.svg';
+    heartIcon3.alt = 'heart-icon-for-likes';
+    userPublishedPostActions.appendChild(heartIcon3);
+
+    container.prepend(userPublishedPost);
+
+    onSnapshot(doc(db, 'post', id), (snapshotDoc) => {
+      const updatedPost = snapshotDoc.data();
+      userPublishedPostText.textContent = updatedPost.text;
+    });
+  }
 
   // Mostrar feed
   window.addEventListener('DOMContentLoaded', async () => {
