+++ src/styles.css
@@ -5,6 +5,7 @@ body {
   margin: 0;
   width: 100%;
   height: 100%;
+  background-color: #FAFAFA;
 }
 
 body,
@@ -820,6 +821,14 @@ main {
   display: flex;
   flex-direction: column;
   margin-left: 10px;
+  width: 100%;
+}
+
+.text-content-upper-row {
+  height: 25px;
+  display: flex;
+  flex-direction: row;
+  padding-right: 5px;
 }
 
 .user-published-post-title {
@@ -827,6 +836,55 @@ main {
   font-weight: 600;
 }
 
+.user-published-post-edit {
+  margin-top: 0px;
+  margin-left: auto;
+  font-size: 25px;
+  display: inline;
+  position: relative;
+  align-self: center;
+  color: #504f4f;
+  cursor: pointer;
+}
+
+/**/
+.edit-post-option-container {
+  border: 1px solid #B6B6B6;
+  padding: 0px;
+  border-radius: 10px;
+  display: none;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  position: absolute;
+  top: 12px;
+  left: -5px;
+  width: 65px;
+  height: fit-content;
+}
+.user-published-post-edit:hover .edit-post-option-container{
+  display: list-item;
+  list-style: none;
+}
+
+.edit-post-option {
+  width: 55px;
+  list-style: none;
+  font-size: 13px;
+  padding: 5px 5px 5px 5px;
+}
+.edit-post-option:hover {
+  border-radius: 8px 8px 0px 0px;
+  background-color:#B6B6B6;
+  color: #FAFAFA;
+}
+.two{
+  border-top: 1px solid #B6B6B6;
+}
+.two:hover{
+  border-radius: 0px 0px 8px 8px;
+}
+/**/
 .user-published-post-text {
   font-size: 13px;
   text-justify: distribute-all-lines;
